Moved the delegate to its own file
