+++ Source/CollectionViewDataSource.swift
@@ -8,27 +8,6 @@
 
 import UIKit
 
-@objc
-public protocol CollectionViewDataSourceDelegate {
-    @objc
-    optional func collectionView(_ collectionView: UICollectionView, canMoveItemAt indexPath: IndexPath) -> Bool
-
-    @objc
-    optional func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell?
-
-    @objc
-    optional func collectionView(_ collectionView: UICollectionView, moveItemAt sourceIndexPath: IndexPath, to destinationIndexPath: IndexPath)
-
-    @objc
-    optional func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int
-
-    @objc
-    optional func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView?
-
-    @objc
-    optional func numberOfSections(in collectionView: UICollectionView) -> Int
-}
-
 /// Provides a configuration object for collection view cells.
 public struct CellConfiguration<T> {
     let presenter: CollectionViewDataSource<T>.CellPresenter?
